complete refactor

+++ preserve_data/load_data.rb
@@ -1,24 +1,52 @@
 require 'json'
-require_relative 'create'
+require_relative './module/create'
 
-class JsonRetrieve
-  def retrieve_games
-    games = []
-    game_data = File.open('./json_data/games.json')
-    JSON.parse(game_data.read).each do |game|
-      new_game = Game.new(game['name'], game['activity'], game['release'], multiplayer: game['mode'])
+class Load
+  def load_game_data
+    game_data = []
+    return item_arr unless File.exist('./json_data/games.json')
+
+    file = File.open('./json_data/games.json')
+    parsed = JSON.parse(file.read)
+    parsed.each do |game|
+      new_game = Game.new(game['activity'], game['release'], game['multiplayer'])
+      new_game.id = game['id']
       create_on_retrieval(game, new_game)
-      games << new_game
+      games_data << new_game
     end
     file.close
-    games
+    game_data
   end
 
-  def retrieve_authors(item_arr)
-    authors = []
-    item_arr&.each do |item|
-      authors << item.author
+  def load_book_data
+    book_data = []
+    return item_arr unless File.exist('./json_data/books.json')
+
+    file = File.open('./json_data/books.json')
+    parsed = JSON.parse(file.read)
+    parsed.each do |book|
+      new_book = Book.new(book['publisher'], book['cover_state'], book['release'])
+      new_book.id = book['id']
+      create_on_retrieval(book, new_book)
+      books_data << new_book
     end
-    authors
+    file.close
+    book_data
+  end
+
+  def load_album_data
+    album_data = []
+    return item_arr unless File.exist('./json_data/albums.json')
+
+    file = File.open('./json_data/albums.json')
+    parsed = JSON.parse(file.read)
+    parsed.each do |album|
+      new_album = MusicAlbum.new(album['release'], album['on_spotify'])
+      new_album.id = album['id']
+      create_on_retrieval(album, new_album)
+      albums_data << new_album
+    end
+    file.close
+    album_data
   end
 end
